user audting

+++ versa-springboot-starters/business-service-boot-starter/src/main/java/io/github/ztoany/versa/infra/springboot/starter/business/service/AbstractSimpleCrudBusinessObjectService.java
@@ -2,11 +2,12 @@ package io.github.ztoany.versa.infra.springboot.starter.business.service;
 
 import io.github.ztoany.versa.infra.business.domain.model.BusinessObject;
 import io.github.ztoany.versa.infra.common.exception.EntityNotFoundException;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Optional;
+
 public abstract class AbstractSimpleCrudBusinessObjectService<E, ID, INPUT, OUTPUT> implements BusinessObjectService<ID, INPUT, OUTPUT> {
-    protected abstract JpaRepository<E, ID> getEntityRepository();
+    protected abstract JpaSpecificationRepository<E, ID> getEntityRepository();
 
     @Override
     @Transactional(rollbackFor = Exception.class)
@@ -23,7 +24,7 @@ public abstract class AbstractSimpleCrudBusinessObjectService<E, ID, INPUT, OUTP
     @Transactional(rollbackFor = Exception.class)
     public OUTPUT update(ID id, INPUT input) {
         var entityRepository = getEntityRepository();
-        var op = entityRepository.findById(id);
+        var op = findEntityById(id);
         var entity = op.orElseThrow(() -> entityNotFoundException(id));
         BusinessObject<E, INPUT> bo = buildBusinessObjectFromEntity(entity);
         beforeUpdate(bo);
@@ -41,23 +42,26 @@ public abstract class AbstractSimpleCrudBusinessObjectService<E, ID, INPUT, OUTP
     @Override
     @Transactional(rollbackFor = Exception.class)
     public OUTPUT getById(ID id) {
-        var entityRepository = getEntityRepository();
-        var op = entityRepository.findById(id);
+        var op = findEntityById(id);
         return op.map(e -> businessObjectToDto(buildBusinessObjectFromEntity(e)))
                 .orElseThrow(() -> entityNotFoundException(id));
     }
 
     protected void deleteByIdInternal(ID id, boolean physicalDelete) {
         var entityRepository = getEntityRepository();
-        var op = entityRepository.findById(id);
-        op.ifPresent(entity -> {
+        var op = findEntityById(id);
+        if(op.isPresent()) {
+            var entity = op.get();
             BusinessObject<E, INPUT> bo = buildBusinessObjectFromEntity(entity);
             bo.onDelete();
-        });
-
-        if(physicalDelete) {
-            entityRepository.deleteById(id);
+            if(physicalDelete) {
+                entityRepository.deleteById(id);
+            }
         }
+    }
+
+    protected Optional<E> findEntityById(ID id) {
+        return getEntityRepository().findById(id);
     }
 
     protected abstract EntityNotFoundException entityNotFoundException(ID id);
